added assets compression support

+++ bundle/jsx/exporters/bannerExporter.jsx
@@ -26,28 +26,35 @@ $.__bodymovin.bm_bannerExporter = (function () {
 		return sourcePath;
 	}
 
-	function createTemplate(config, destinationPath) {
+	function getSizes(bannerConfig) {
+		return {
+			width: bannerConfig.use_original_sizes ? bannerConfig.original_width : bannerConfig.width,
+			height: bannerConfig.use_original_sizes ? bannerConfig.original_height : bannerConfig.height,
+		}
+	}
+
+	function createTemplate(config, filePathName, savingFolder) {
 		var bannerConfig = config.banner
 		var templateData = bm_downloadManager.getTemplateData()
-		var filePathName = destinationPath.substr(destinationPath.lastIndexOf('/') + 1);
+		var sizes = getSizes(bannerConfig)
 
 		templateData = templateData
-		.replace(/__CONTENT_WIDTH__/g, bannerConfig.width)
-		.replace(/__CONTENT_HEIGHT__/g, bannerConfig.height)
+		.replace(/__CONTENT_WIDTH__/g, sizes.width)
+		.replace(/__CONTENT_HEIGHT__/g, sizes.height)
 		.replace(/__LOTTIE_RENDERER__/g, bannerConfig.lottie_renderer)
 		.replace(/__CLICK_TAG__/g, bannerConfig.click_tag)
 		.replace(/__LOTTIE_SOURCE__/g, getLottiePath(bannerConfig))
 		.replace(/__FILE_PATH__/g, filePathName)
 
-		var htmlDestinationPath = destinationPath.replace(filePathName,'index.html');
-		var dataFile = new File(htmlDestinationPath);
+		var dataFile = new File(savingFolder.fullName + '/index.html');
 		dataFile.open('w', 'TEXT', '????');
 		dataFile.encoding = 'UTF-8';
 		dataFile.write(templateData);
 		dataFile.close();
+
 	}
 
-	function includeAdditionalFiles(config, destinationPath) {
+	function includeAdditionalFiles(config, savingFolder) {
 		if (!bm_projectManager) {
 			bm_projectManager = $.__bodymovin.bm_projectManager;
 		}
@@ -63,9 +70,6 @@ $.__bodymovin.bm_bannerExporter = (function () {
 			// var file = bm_projectManager.getFile('/assets/player/lottie.js')
 			var file = bm_projectManager.getFile('/assets/player/' + sourcePath)
 
-			bm_eventDispatcher.log('exists: ' + file.exists)
-			bm_eventDispatcher.log('absoluteURI: ' + file.absoluteURI)
-
 
 			/*var i = 0, len = lottiePaths.length;
 			while ( i < len) {
@@ -75,25 +79,47 @@ $.__bodymovin.bm_bannerExporter = (function () {
 				i += 1;
 			}*/
 
-			var filePathName = destinationPath.substr(destinationPath.lastIndexOf('/') + 1);
-			var playerDestinationPath = destinationPath.replace(filePathName, sourcePath);
-			var dataFile = new File(playerDestinationPath);
-			// dataFile.open('w', 'TEXT', '????');
-			// dataFile.encoding = 'UTF-8';
+			var dataFile = new File(savingFolder.fullName + '/' + sourcePath);
 			
 			file.copy( dataFile )
 		}
 	}
 
 	function save(stringData, destinationPath, config, callback) {
-		createTemplate(config, destinationPath)
-		includeAdditionalFiles(config, destinationPath)
-		var dataFile = new File(destinationPath);
-		dataFile.open('w', 'TEXT', '????');
-		dataFile.encoding = 'UTF-8';
-		dataFile.write(stringData);
-		dataFile.close();
 
+		var bannerConfig = config.banner
+		var savingFolder
+		var destinationPathFolder = destinationPath.substr(0, destinationPath.lastIndexOf('/') + 1);
+		var destinationFolder = new Folder(destinationPathFolder)
+		if (bannerConfig.zip_files) {
+			var tempFolderName = 'tmp_' + Math.random().toString().substr(2)
+			savingFolder = new Folder(Folder.temp.path + '/LottieTemp/' + tempFolderName)
+		} else {
+			savingFolder = destinationFolder
+		}
+
+		if (savingFolder.exists || savingFolder.create()) {
+
+			var filePathName = destinationPath.substr(destinationPath.lastIndexOf('/') + 1);
+			var jsonPathName = filePathName
+			if (bannerConfig.zip_files) {
+				jsonPathName = jsonPathName.replace('.zip', '.json')
+			}
+
+			createTemplate(config, jsonPathName, savingFolder)
+			includeAdditionalFiles(config, savingFolder)
+			var dataFile = new File(savingFolder.fullName + '/' + jsonPathName);
+			dataFile.open('w', 'TEXT', '????');
+			dataFile.encoding = 'UTF-8';
+			dataFile.write(stringData);
+			dataFile.close();
+
+			
+			if (bannerConfig.zip_files) {
+				bm_eventDispatcher.sendEvent('bm:zip:banner', {destinationPath: destinationFolder.fsName + '/' + filePathName, folderPath: savingFolder.fsName});
+			}
+			
+		}
 		callback();
 	}
